+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2015-09-03
+
+* Add `Close` method to melody instance.
+
 ### 2015-06-10
 
 * Support for binary messages.